+++ Runtime/Scripts/DracoMeshLoader.cs
@@ -18,19 +18,55 @@ using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 using Unity.Jobs;
 using UnityEngine;
+using UnityEngine.Profiling;
+using UnityEngine.Rendering;
 
 namespace Draco {
 
     public class DracoMeshLoader
     {
+        public struct DecodeResult {
+            /// <summary>
+            /// True if the decoding was successful
+            /// </summary>
+            public bool success;
+            
+            /// <summary>
+            /// True, if the normals were marked required, but not present in Draco mesh.
+            /// They have to get calculated.
+            /// </summary>
+            public bool calculateNormals;
+        }
+
+        public const MeshUpdateFlags defaultMeshUpdateFlags = MeshUpdateFlags.DontRecalculateBounds | MeshUpdateFlags.DontValidateIndices | MeshUpdateFlags.DontNotifyMeshUsers | MeshUpdateFlags.DontResetBoneBounds;
+        
         /// <summary>
         /// Decodes a Draco mesh
         /// </summary>
         /// <param name="encodedData">Compressed Draco data</param>
         /// <returns>Unity Mesh or null in case of errors</returns>
-        public async Task<Mesh> ConvertDracoMeshToUnity(NativeArray<byte> encodedData) {
+        public async Task<Mesh> ConvertDracoMeshToUnity(NativeArray<byte> encodedData, bool requireNormals = false, bool requireTangents = false) {
             var encodedDataPtr = GetUnsafeReadOnlyIntPtr(encodedData);
-            return await ConvertDracoMeshToUnity(encodedDataPtr, encodedData.Length);
+#if UNITY_2020_2_OR_NEWER
+            var meshDataArray = Mesh.AllocateWritableMeshData(1); 
+            var mesh = meshDataArray[0];
+            var result = await ConvertDracoMeshToUnity(mesh, encodedDataPtr, encodedData.Length, requireNormals, requireTangents);
+            if (!result.success) {
+                meshDataArray.Dispose();
+                return null;
+            }
+            var unityMesh = new Mesh();
+            Mesh.ApplyAndDisposeWritableMeshData(meshDataArray,unityMesh,defaultMeshUpdateFlags);
+            if (result.calculateNormals) {
+                unityMesh.RecalculateNormals();
+            }
+            if (requireTangents) {
+                unityMesh.RecalculateTangents();
+            }
+            return unityMesh;
+#else
+            return await ConvertDracoMeshToUnity(encodedDataPtr, encodedData.Length, requireNormals, requireTangents);
+#endif
         }
 
         /// <summary>
@@ -38,27 +74,109 @@ namespace Draco {
         /// </summary>
         /// <param name="encodedData">Compressed Draco data</param>
         /// <returns>Unity Mesh or null in case of errors</returns>
-        public async Task<Mesh> ConvertDracoMeshToUnity(byte[] encodedData) {
+        public async Task<Mesh> ConvertDracoMeshToUnity(byte[] encodedData, bool requireNormals = false, bool requireTangents = false) {
             var encodedDataPtr = PinGCArrayAndGetDataAddress(encodedData, out var gcHandle);
-            var result = await ConvertDracoMeshToUnity(encodedDataPtr, encodedData.Length);
+#if UNITY_2020_2_OR_NEWER
+            using (var meshDataArray = Mesh.AllocateWritableMeshData(1)) {
+                var mesh = meshDataArray[0];
+                var result = await ConvertDracoMeshToUnity(mesh, encodedDataPtr, encodedData.Length, requireNormals, requireTangents);
+                UnsafeUtility.ReleaseGCObject(gcHandle);
+                if (!result.success) return null;
+                var unityMesh = new Mesh();
+                Mesh.ApplyAndDisposeWritableMeshData(meshDataArray,unityMesh,defaultMeshUpdateFlags);
+                if (result.calculateNormals) {
+                    unityMesh.RecalculateNormals();
+                }
+                if (requireTangents) {
+                    unityMesh.RecalculateTangents();
+                }
+                return unityMesh;
+            }
+#else
+            var result = await ConvertDracoMeshToUnity(encodedDataPtr, encodedData.Length, requireNormals, requireTangents);
+            UnsafeUtility.ReleaseGCObject(gcHandle);
+            return result;
+#endif
+        }
+
+#if UNITY_2020_2_OR_NEWER
+        public async Task<DecodeResult> ConvertDracoMeshToUnity(Mesh.MeshData mesh, byte[] encodedData, bool requireNormals = false, bool requireTangents = false) {
+            var encodedDataPtr = PinGCArrayAndGetDataAddress(encodedData, out var gcHandle);
+            var result = await ConvertDracoMeshToUnity(mesh, encodedDataPtr, encodedData.Length, requireNormals, requireTangents);
             UnsafeUtility.ReleaseGCObject(gcHandle);
             return result;
         }
-
-        async Task<Mesh> ConvertDracoMeshToUnity(IntPtr encodedData, int size) {
+        
+        public async Task<DecodeResult> ConvertDracoMeshToUnity(Mesh.MeshData mesh, NativeArray<byte> encodedData, bool requireNormals = false, bool requireTangents = false) {
+            var encodedDataPtr = GetUnsafeReadOnlyIntPtr(encodedData);
+            return await ConvertDracoMeshToUnity(mesh, encodedDataPtr, encodedData.Length, requireNormals, requireTangents);
+        }
+#endif
+        
+#if UNITY_2020_2_OR_NEWER
+        async Task<DecodeResult> ConvertDracoMeshToUnity(
+            Mesh.MeshData mesh,
+            IntPtr encodedData,
+            int size,
+            bool requireNormals,
+            bool requireTangents
+        )
+#else
+        async Task<Mesh> ConvertDracoMeshToUnity(IntPtr encodedData, int size, bool requireNormals, bool requireTangents)
+#endif
+        {
+#if UNITY_2020_2_OR_NEWER
+            var dracoNative = new DracoNative(mesh);
+            var result = new DecodeResult();
+#else
             var dracoNative = new DracoNative();
+#endif
             await WaitForJobHandle(dracoNative.Init(encodedData, size));
             if (dracoNative.ErrorOccured()) {
+#if UNITY_2020_2_OR_NEWER
+                return result;
+#else
                 return null;
+#endif
             }
-            dracoNative.CreateMesh();
+            if (!requireNormals && requireTangents) {
+                // Sanity check: We need normals to calculate tangents
+                requireNormals = true;
+            }
+#if UNITY_2020_2_OR_NEWER
+            dracoNative.CreateMesh(out result.calculateNormals, requireNormals,requireTangents);
+#else
+            dracoNative.CreateMesh(out var calculateNormals, requireNormals,requireTangents);
+#endif      
             await WaitForJobHandle(dracoNative.DecodeVertexData());
             var error = dracoNative.ErrorOccured();
             dracoNative.DisposeDracoMesh();
             if (error) {
+#if UNITY_2020_2_OR_NEWER
+                return result;
+#else
                 return null;
+#endif
             }
-            return dracoNative.PopulateMeshData();
+
+#if !UNITY_2020_2_OR_NEWER
+            var result = dracoNative.PopulateMeshData();
+            if (calculateNormals) {
+                // TODO: Consider doing this in a threaded Job
+                Profiler.BeginSample("RecalculateNormals");
+                result.RecalculateNormals();
+                Profiler.EndSample();
+            }
+            if (requireTangents) {
+                // TODO: Consider doing this in a threaded Job
+                Profiler.BeginSample("RecalculateTangents");
+                result.RecalculateTangents();
+                Profiler.EndSample();
+            }
+#else
+            result.success = dracoNative.PopulateMeshData();
+#endif
+            return result;
         }
 
         static async Task WaitForJobHandle(JobHandle jobHandle) {