+++ Tests/Runtime/DracoRuntimeTests.cs
@@ -50,7 +50,43 @@ namespace Draco.Tests
             yield return RunTest(url);
         }
 
-        IEnumerator RunTest(string url) {
+        [UnityTest]
+        [UseDracoTestFileCase(new[] {
+            "https://raw.githubusercontent.com/google/draco/master/testdata/car.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att.obj.edgebreaker.cl10.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att.obj.edgebreaker.cl4.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att.obj.sequential.cl3.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att_sub_o_2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att_sub_o_no_metadata.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.edgebreaker.1.2.0.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.edgebreaker.cl10.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.edgebreaker.cl4.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.sequential.1.2.0.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.sequential.cl3.2.2.drc",
+        })]
+        public IEnumerator LoadDracoOfficialTestDataNormals(string url) {
+            yield return RunTest(url,true);
+        }
+        
+        [UnityTest]
+        [UseDracoTestFileCase(new[] {
+            "https://raw.githubusercontent.com/google/draco/master/testdata/car.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att.obj.edgebreaker.cl10.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att.obj.edgebreaker.cl4.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att.obj.sequential.cl3.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att_sub_o_2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/cube_att_sub_o_no_metadata.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.edgebreaker.1.2.0.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.edgebreaker.cl10.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.edgebreaker.cl4.2.2.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.sequential.1.2.0.drc",
+            "https://raw.githubusercontent.com/google/draco/master/testdata/test_nm.obj.sequential.cl3.2.2.drc",
+        })]
+        public IEnumerator LoadDracoOfficialTestDataNormalsTangents(string url) {
+            yield return RunTest(url,true, true);
+        }
+        
+        IEnumerator RunTest(string url, bool requireNormals = false, bool requireTangents = false) {
             var webRequest = UnityWebRequest.Get(url);
             yield return webRequest.SendWebRequest();
             if(!string.IsNullOrEmpty(webRequest.error)) {
@@ -59,23 +95,23 @@ namespace Draco.Tests
             }
 
             var data = new NativeArray<byte>(webRequest.downloadHandler.data,Allocator.Persistent);
-            var go = new GameObject("DracoBenchmark");
-
-            var task = LoadBatch(1, data);
+            
+            var task = LoadBatch(1, data, requireNormals, requireTangents);
             while (!task.IsCompleted) {
                 yield return null;
             }
+            Assert.IsNull(task.Exception);
             data.Dispose();
         }
         
-        async Task LoadBatch(int quantity, NativeArray<byte> data) {
+        async Task LoadBatch(int quantity, NativeArray<byte> data, bool requireNormals = false, bool requireTangents = false) {
 
             var tasks = new List<Task<Mesh>>(quantity);
         
             for (var i = 0; i < quantity; i++)
             {
                 DracoMeshLoader dracoLoader = new DracoMeshLoader();
-                var task = dracoLoader.ConvertDracoMeshToUnity(data);
+                var task = dracoLoader.ConvertDracoMeshToUnity(data,requireNormals,requireTangents);
                 tasks.Add(task);
             }
 
@@ -85,6 +121,16 @@ namespace Draco.Tests
                 var mesh = await task;
                 if (mesh == null) {
                     Debug.LogError("Loading mesh failed");
+                }
+                else {
+                    if (requireNormals) {
+                        var normals = mesh.normals;
+                        Assert.Greater(normals.Length,0);
+                    }
+                    if (requireTangents) {
+                        var tangents = mesh.tangents;
+                        Assert.Greater(tangents.Length,0);
+                    }
                 }
             }
             await Task.Yield();